Change script to only remove the grandma voice
Changed the script to only remove the grandma voice. We could probably remove grandpa too but we should leave the others and see if they cause issues.

Also changing the folder they are moved to. Don't want to drop it in home where the user can see and delete it. Now it is in the 'v!' folder. This hides it from the user. Might not be the best folder though.